🚨 Fix Detekt and Ktlint issues
Wow! A lot of things showed up after the implementation!
