+++ app/src/main/java/com/escodro/alkaa/presentation/home/Home.kt
@@ -30,6 +30,7 @@ import com.escodro.task.presentation.list.TaskListSection
 /**
  * Alkaa Home screen.
  */
+@Suppress("LongParameterList")
 @Composable
 fun Home(
     onTaskClick: (Long) -> Unit,
